Pointless Button V2 - Wiring Schematic
This is the current wiring schematic. The ESP32 "header" on the right side is a general labeling and does not match any board, you will have to determine what pins you are connecting to when you build yours. Same goes for the 18650 Cells, you can use different cells for your pack, just adjust accordingly.