Add files via upload
I did not put shrink wrap all the way up on the pins the regulator is connected to for easy replacement if I burn it out or need to change it for a different one. I did cover it with a strip of coroplast tape so it wont melt or ground out to something.